requirements.txt
fastapi==0.88.0
httpx==0.24.1
python-jose==3.3.0
pyjwt==2.4.0
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.5
pytest
pytest-asyncio
pytest-cov